new structure. Old functionalities seem ok
