+++ marshaltools/__init__.py
@@ -1,10 +1,16 @@
 from __future__ import absolute_import
 import os
 from .version import __VERSION__ as __version__
-from .marshaltools import *
+#from .marshaltools import *
+
 from .surveyfields import SurveyFields, ZTFFields
+from .BaseTable import BaseTable
+from .MarshalLightcurve import MarshalLightcurve
+from .ProgramList import ProgramList
+
 from .filters import load_filters
 load_filters()
+
 
 here = __file__
 # basedir = os.path.split(here)[0]
