+++ tests/test0.py
@@ -1,0 +1,2 @@
+import marshaltools
+prog = marshaltools.ProgramList('Cosmology')
