+++ tests/test1.py
@@ -1,0 +1,13 @@
+import marshaltools
+
+
+prog = marshaltools.ProgramList('Cosmology')
+
+
+t = prog.table
+name = t['name'][1]
+print (t)
+
+print (name)
+lc = prog.get_lightcurve(name)
+
